+++ CHANGES.rst
@@ -31,6 +31,7 @@ ROMMover
 General
 ~~~~~~~
 
+- Added new "download names", that can map a ROM to a different name
 - Fix crash when version formatted like 'v.21', and added test
 - Updated regex to move Kickstarter, The Retro Room to modern versions
 - Added "Red Art Games" and "Metal Gear Solid Collection" to modern versions in regex